CRUD finished

+++ app/Http/Controllers/QuestionController.php
@@ -17,39 +17,37 @@ class QuestionController extends Controller
     }
     public function indexRandom($signature)
     {
-        if ($signature === "all"){
-        $questions = DB::table('questions')
-                ->orderByRaw("RANDOM()")
-                ->take(3)
-                ->get();
+        if ($signature==0){
+            $questions = Question::inRandomOrder()
+            ->take(3)
+            ->get();
+            return response()->json($questions, 200);
+
         }
-        if ($signature === "Math"){
-            $questions = DB::table('questions')
-                    ->orderByRaw("RANDOM()")
+        if ($signature == 1){
+            $questions = Question::inRandomOrder()
                     ->take(3)
-                    ->where('competencies_id', '=', '1')
+                    ->where('competencies_id', '=', $signature)
                     ->get();
+            return response()->json($questions, 200);
             }
-            if ($signature === "Language"){
-                $questions = DB::table('questions')
-                        ->orderByRaw("RANDOM()")
+            if ($signature == 2){
+            $questions = Question::inRandomOrder()
                         ->take(3)
-                        ->where('competencies_id', '=', '2')
+                        ->where('competencies_id', '=', $signature)
                         ->get();
+            return response()->json($questions, 200);
                 }
-            if ($signature === "English"){
-                $questions = DB::table('questions')
-                        ->orderByRaw("RANDOM()")
-                        ->take(3)
-                        ->where('competencies_id', '=', '3')
-                        ->get();
-                }
-        return response()->json($questions, 200);
+
+
     }
 
     public function destroy($id)
     {
         $question = Question::find($id);
+        if ($question == null){
+            return response()->json("Esa pregunta no existe", 200);
+        }
         $question->delete();
     }
     public function show($id)
@@ -58,16 +56,23 @@ class QuestionController extends Controller
         $question = [];
 
         $question[0] = Question::find($id);
+        if ($question[0] == null) {
+            return response()->json("Esa pregunta no existe", 200);
+        }
         return response()->json($question, 200);
     }
     public function store(Request $request)
     {
         $question = request()->except('_token');
-        question::create($question);
+        Question::create($question);
     }
     public function update(Request $request, $id)
     {
         $question = request()->except(['_token', '_method']);
+        $questionTest = Question::find($id);
+        if ($questionTest == null) {
+            return response()->json("Esa pregunta no existe", 200);
+        }
         Question::where('id', '=', $id)->update($question);
 
     }
